+++ Tests/VCoreTests/Tests/Extensions/Foundation/CollectionFirstAndLastElementTests.swift
@@ -1,0 +1,49 @@
+//
+//  CollectionFirstAndLastElementTests.swift
+//  VCoreTests
+//
+//  Created by Vakhtang Kontridze on 29.11.22.
+//
+
+import XCTest
+@testable import VCore
+
+// MARK: - Tests
+final class CollectionFirstAndLastElementTests: XCTestCase {
+    // MARK: Test Data
+    private let array: [any P] = [
+        S1(value: 1),
+        S1(value: 2),
+        S2(value: 3),
+        S2(value: 4),
+        S1(value: 5),
+        S1(value: 6)
+    ]
+    
+    // MARK: Tests - First
+    func testFirst() {
+        XCTAssertEqual(array.firstElement(ofType: S1.self)?.value, 1)
+    }
+
+    func testFirstPredicate() {
+        XCTAssertEqual(array.firstElement(ofType: S1.self, where: { $0.value > 1 })?.value, 2)
+    }
+
+    // MARK: Tests - Last
+    func testLast() {
+        XCTAssertEqual(array.lastElement(ofType: S1.self)?.value, 6)
+    }
+
+    func testLastPredicate() {
+        XCTAssertEqual(array.lastElement(ofType: S1.self, where: { $0.value < 6 })?.value, 5)
+    }
+}
+
+// MARK: - Test Data
+/*private*/ protocol P {
+    var value: Int { get }
+}
+
+/*private*/ struct S1: P { let value: Int }
+
+/*private*/ struct S2: P { let value: Int }